Make Horizon Dashboard settings operator overrideable.
We want to allow operators to override Horizon's default
settings. This involves moving local_settings.d out of the read only
snap filesystem, and into $SNAP_COMMON. This is a little bit tricky.

First, we edit settings.py and local_settings.py as we're building the
snap, to include a LOCAL_PATH in $SNAP_COMMON.

Then, we add a template with the rest of our default overrides,
and write it out to $SNAP_COMMON/horizon/local_settings.d

Finally we tweak our tests so that we can give our overrides a
spin. As a bonus, this makes running test_horizonglogin.py a lot
easier to run in our multipass testing scenario!

Also tweaked the tox -e basic test so that it can successfully run
selenium tests as well.

Change-Id: Ic0ce18cfa1b97a93191da749095d8aa2270d5aeb